chore: address /code-review feedback — dead code, narrow catch, helper extraction
- Remove unimported test-helper waitForScanStatus (was speculative)
- Remove unread lastTickAt field on ThrottleScheduler (write-only state)
- Document why moveSameDrive uses row.sha256 as post_hash (rename
  atomicity preserves bytes; cross-drive path differs because pipeline
  copies can corrupt mid-stream)
- Narrow parseJsonBody catch to SyntaxError so non-parse exceptions
  propagate to onError instead of being mis-mapped to invalid-json
- Extract makeCapturingLogger to test-helpers; metadata-image.test.ts
  and metadata-video.test.ts now import the shared helper
