+++ packages/engine/src/api/server.ts
@@ -65,8 +65,9 @@ const CACHE_CONTROL_MAX_AGE = 'max-age=300';
 async function parseJsonBody<T>(c: Context): Promise<T | null> {
   try {
     return (await c.req.json()) as T;
-  } catch {
-    return null;
+  } catch (err) {
+    if (err instanceof SyntaxError) return null;
+    throw err;
   }
 }
