+++ packages/engine/src/test-helpers/log.ts
@@ -10,3 +10,19 @@ export function silentLogger(): Logger {
   };
   return log;
 }
+
+export function makeCapturingLogger(): {
+  logger: Logger;
+  warns: Array<{ msg: string; fields: Record<string, unknown> }>;
+} {
+  const warns: Array<{ msg: string; fields: Record<string, unknown> }> = [];
+  const noop = () => {};
+  const logger: Logger = {
+    debug: noop,
+    info: noop,
+    warn: (msg, fields) => warns.push({ msg, fields: fields ?? {} }),
+    error: noop,
+    child: () => logger,
+  };
+  return { logger, warns };
+}
